backend/requirements.txt
fastapi
uvicorn
neo4j
google-generativeai
wordninja
sentence-transformers
faiss-cpu
pdfplumber
google-genai
pandas
PyPDF2
nltk
ollama
python-dotenv